backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
firebase-admin==6.5.0
python-dotenv==1.0.1
pandas==2.2.2
scikit-learn==1.5.1
numpy==1.26.4
africastalking==1.2.7
geopy==2.4.1
gunicorn==22.0.0
